final fixes

+++ Investment-Service/src/main/java/ru/tuganov/dto/InstrumentDto.java
@@ -1,4 +1,4 @@
 package ru.tuganov.dto;
 
-public record InstrumentDto(String name, String figi, long sellPrice, long buyPrice) {
+public record InstrumentDto(String name, String figi, double sellPrice, double buyPrice) {
 }
